hot fix

+++ src/lexer/token_utils.c
@@ -6,7 +6,7 @@
 /*   By: yabarhda <yabarhda@student.1337.ma>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/07/28 21:53:48 by yabarhda          #+#    #+#             */
-/*   Updated: 2025/07/29 12:57:24 by yabarhda         ###   ########.fr       */
+/*   Updated: 2025/07/30 04:58:42 by yabarhda         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -49,10 +49,10 @@ void free_tokens(t_token *tokens)
         tmp = tokens;
         tokens = tokens->next;
 		if (tmp->type == TOKEN_HEREDOC)
-			unlink(tmp->next->value);
-        // if (tmp->value && tmp->type != TOKEN_HEREDOC)
-        //     free(tmp->value);
-        // free(tmp);
+		{
+			if (tmp->next)
+				unlink(tmp->next->value);
+		}
     }
 }
